pyproject.toml
[tool.poetry]
name = "pypi-package-stats"
version = "0.1.0"
description = "For send BigQuery Data to DataDog"
authors = ["Ivanildo Barauna de Souza Junior <ivanildo.jnr@outlook.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-secret-manager = "^2.20.2"
google-cloud-bigquery = "^3.27.0"
datadog = "^0.50.1"
pandas = "^2.2.3"
db-dtypes = "^1.3.1"
black = "^24.10.0"
datadog-api-client = "^2.30.0"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
coverage = "^7.6.4"
tqdm = "^4.66.5"
pydantic = "^2.10.3"
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-exporter-otlp = "^1.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
coverage = "^7.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
